testing k_resolution

+++ vscworkflows/setup/sets.py
@@ -85,7 +85,7 @@ class BulkStaticSet(DictSet):
         if "k_resolution" in settings:
             # Use k_resolution to calculate kpoints
             k_kpoint_resolution = settings["k_resolution"]
-            kpt_divisions = [np.floor(1 / l / k_kpoint_resolution) for l in
+            kpt_divisions = [int(np.floor(1 / l / k_kpoint_resolution)) for l in
                              self.structure.lattice.lengths]
 
             return Kpoints.gamma_automatic(kpts=kpt_divisions)
